Specify run time in minutes
When grouping jobs in Snakemake, it tries to take the max run time for
jobs that will run in parallel and the sum of run times for jobs that
run in series. If the run time is a string, however, it will not be able
to do that, and will thus require them to be identical. This should fix
that issue.