refactor: restructure Twist Solid config files (#6)
* New structure for twist-solid-0.13.0

The idea is that this is adapted for other pipelines/versions too. Right
now I'm not sure if it's worth porting previous versions, or if I should
just remove them.

* Remove old twist-solid-0.13.0 config

* Correct config path

* Don't pass environment to apptainer

* Move slurm status script to more central location

This script is general for anything that is being submitted to slurm, at
least when using snakemake. Now we can avoid having to edit it in a lot
of places whenever anything breaks, or we want new features.

* Remove config for old versions of twist solid

These are more or less obsolete by now, and if we for some reason want
to go back, then it will likely be easier to recreate them from scratch.
I don't see this happening though.

* Add simple validation for snakemake profiles

feat: add validation script for Snakemake profiles
feat: remove Twist Solid config files for versions <0.13.0