feat: add configuration files for nf-core/raredisease (#9)
* Initial config for nf-core raredisease

* Move raredisease-specific profiles

* Add some documentation on usage

* Add interval lists for GRCh38 iGenomes reference

* Flesh out the vcfanno config

* More resources for deepvariant and vcfanno

* Add VEP plugins

This allows VEP to actually run as it is supposed to. There are still
discrepancies between the annotations that are added and the ones that
genmod expects given the current rank model, so things will have to be
updated.

* Fix `--cleanenv` argument for `apptainer exec`

* Structure institutional configs as in nf-core/configs

* Custom VEP config

* Set up separate rank model for raredisease

* Fix swegen and most severe consequence

* Add REVEL annotations

* Add CLINVAR annotations

* Remove SPIDEX from the rank model

I have only been able to find scores for GRCh37, and doing liftover on
SNPs is not recommended. An alternative approach for converting these
could be to find rsIDs for as many variants as possible and do the
conversion that way, but I will exclude it for now.

* Add dbNSFP annotations

* Use custom_config_base instead of hard coding path

* Disable VEP gene filtering for now

* Restructure nf-core configs

Trying to replicate the config file structure of nf-core proved to be
more messy than anticipated. This new structure should be much more
simple to manage, and also much easier to know where something has been
defined.