feat: add plumberfile (#19)
* Add plumber file for nf-core/raredisease

* Add twist-solid to plumber file

* Remove unused keys for twist solid

* Update twist solid name to be pipeline repo

* Don't use relative includes

This will break when using it with plumber. Instead, any additional
config files should be supplied on the command line.

* Fix asset paths

This allows for the config repo to be used standalone, as well as from
plumber. The exception that is raised only mentions
`NEXTFLOW_CONFIG_HOME` since if it's being run with plumber, then the
environment variable is set for you, and you shouldn't have to even know
about it.

* Add plumberfile documentation to readme