feat: add config for Twist Solid v0.23.0 (#28)
* Add Twist Solid 0.23.0 to plumberfile

* Add config files

* Add executor for snakemake pipelines

* Fix profile paths

* Fix versions in paths

* Fix resource path

* Remove hotspot report CRC definition

* Add MultiQC config for general report

* Adapt plumberfile to new structure

* Use version as specified in the original repo

* Move general resources to assets

This is not used as a config file, so it shouldn't be treated as one

* Add pipeline home and config to environment

* Switch to drmaa for pipeline execution

This will enable cluster execution without having to rely on a separate
script for checking job status on the cluster. As far as I can tell
everything else is equivalent to how it worked before.

* Update readme

* Update Twist Solid v0.14.0

* Rename PLUMBER_ASSETS_PATH -> PLUMBER_PIPELINE_ASSETS